api_with_restrictions/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter==23.3
flake8==6.1.0
python-dotenv==1.0.0
gunicorn==21.2.0